+++ cli/commands/metadata_apply.go
@@ -1,21 +1,12 @@
 package commands
 
 import (
-	"bufio"
-	"bytes"
-	"encoding/json"
-	"fmt"
-	"io"
-	"io/ioutil"
-	"os"
+  "fmt"
+  "os"
 
-	"github.com/goccy/go-yaml"
-
-	"github.com/hasura/graphql-engine/cli/internal/hasura"
-
-	"github.com/hasura/graphql-engine/cli"
-	"github.com/hasura/graphql-engine/cli/internal/metadataobject"
-	"github.com/spf13/cobra"
+  "github.com/hasura/graphql-engine/cli"
+  "github.com/hasura/graphql-engine/cli/internal/metadataobject"
+  "github.com/spf13/cobra"
 )
 
 func newMetadataApplyCmd(ec *cli.ExecutionContext) *cobra.Command {
@@ -68,40 +59,6 @@ type MetadataApplyOptions struct {
 
 func (o *MetadataApplyOptions) Run() error {
 	metadataHandler := metadataobject.NewHandlerFromEC(o.EC)
-	// check for any input from pipe
-	info, err := os.Stdin.Stat()
-	if err != nil {
-		o.EC.Logger.Debug("looks look we didn't find a stdin")
-	}
-	if err == nil {
-		if (info.Mode() & os.ModeCharDevice) == 0 {
-			reader := bufio.NewReader(os.Stdin)
-			o.EC.Logger.Debug("applying metadata from pipe")
-			metadataJSON, err := getMetadataJSON(ec, reader)
-			if err != nil {
-				return err
-			}
-			if !o.DryRun {
-				if err := replaceMetadata(ec, metadataJSON); err != nil {
-					return errorApplyingMetadata(err)
-				}
-				if len(o.rawOutput) != 0 {
-					// if not  a dry run after applying metadata, get a dump from server and print into stdout
-					return getMetadataFromServerAndWriteToStdoutByFormat(o.EC, rawOutputFormat(o.rawOutput))
-				}
-				return nil
-			} else {
-				// on a dry run write the input to stdout by format
-				outputFormat := rawOutputFormatJSON
-				if len(o.rawOutput) > 0 {
-					outputFormat = rawOutputFormat(o.rawOutput)
-				}
-				return writeByOutputFormat(os.Stdout, metadataJSON, outputFormat)
-			}
-			return nil
-		}
-	}
-
 	if !o.DryRun {
 		o.EC.Spin("Applying metadata...")
 		if o.EC.Config.Version == cli.V2 {
@@ -155,59 +112,8 @@ func (o *MetadataApplyOptions) Run() error {
 	return nil
 }
 
-// get metadata from reader is it is in JSON/YAML format
-// returns an error otherwise
-func getMetadataJSON(ec *cli.ExecutionContext, reader io.Reader) ([]byte, error) {
-	maybeMetadata, err := ioutil.ReadAll(reader)
-	if err != nil {
-		return nil, fmt.Errorf("reading input from os.Stdin: %w", err)
-	}
-	if len(maybeMetadata) == 0 {
-		return nil, fmt.Errorf("found 0 bytes from input")
-	}
-	var metadata []byte
-	if isJSON(maybeMetadata) {
-		metadata = maybeMetadata
-		ec.Logger.Debug("found JSON metadata from pipe")
-	} else if isYAML(maybeMetadata) {
-		ec.Logger.Debug("found YAML metadata from pipe")
-		yamlMetadata, err := yaml.YAMLToJSON(maybeMetadata)
-		if err != nil {
-			return nil, fmt.Errorf("error decoding input as json: %w", err)
-		}
-		metadata = yamlMetadata
-	} else {
-		return nil, fmt.Errorf("cannot determine format from input, accepted input formats: [json,yaml]")
-	}
-	return metadata, nil
-}
-
 func errorApplyingMetadata(err error) error {
 	// a helper function to have consistent error messages for errors
 	// when applying metadata
 	return fmt.Errorf("error applying metadata \n%w", err)
-}
-
-func replaceMetadata(ec *cli.ExecutionContext, metadataJSON []byte) error {
-	if ec.Config.Version == cli.V2 {
-		if _, err := cli.GetCommonMetadataOps(ec).ReplaceMetadata(bytes.NewReader(metadataJSON)); err != nil {
-			return err
-		}
-	} else {
-		var v interface{}
-		if err := json.Unmarshal(metadataJSON, &v); err != nil {
-			return err
-		}
-		r, err := ec.APIClient.V1Metadata.V2ReplaceMetadata(hasura.V2ReplaceMetadataArgs{
-			AllowInconsistentMetadata: true,
-			Metadata:                  v,
-		})
-		if err != nil {
-			return err
-		}
-		if !r.IsConsistent {
-			ec.Logger.Warn("Metadata is inconsistent")
-		}
-	}
-	return nil
 }