+++ cli/commands/metadata_apply_test.go
@@ -2,15 +2,12 @@ package commands
 
 import (
 	"fmt"
-	"os"
-	"os/exec"
-	"path/filepath"
-	"strings"
-
 	"github.com/hasura/graphql-engine/cli/internal/testutil"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	. "github.com/onsi/gomega/gexec"
+	"os"
+	"path/filepath"
 )
 
 var _ = Describe("metadata_apply", func() {
@@ -57,49 +54,6 @@ var _ = Describe("metadata_apply", func() {
 			})
 			Eventually(session, 60*40).Should(Exit(0))
 			Eventually(isJSON(session.Wait().Out.Contents())).Should(BeTrue())
-		})
-
-		Context("metadata dry run ", func() {
-			args := strings.Join([]string{
-				testutil.CLIBinaryPath, "metadata", "export", "-o", "json",
-				"|", "jq", `'.sources[0].configuration.connection_info.pool_settings.max_connections = 51'`,
-				"|", testutil.CLIBinaryPath, "md", "apply", "--dry-run",
-			}, " ")
-			cmd := exec.Command("bash", "-c", args)
-			cmd.Dir = dirName
-			session, err := Start(
-				cmd,
-				NewPrefixedWriter(testutil.DebugOutPrefix, GinkgoWriter),
-				NewPrefixedWriter(testutil.DebugErrPrefix, GinkgoWriter),
-			)
-			Expect(err).To(BeNil())
-			Eventually(session, 60*40).Should(Exit(0))
-			Eventually(string(session.Wait().Out.Contents())).Should(ContainSubstring("\"max_connections\": 51\n"))
-		})
-		Context("metadata apply metadata from pipe", func() {
-			args := strings.Join([]string{
-				testutil.CLIBinaryPath, "metadata", "export", "-o", "json",
-				"|", "jq", `'.sources[0].configuration.connection_info.pool_settings.max_connections = 51'`,
-				"|", testutil.CLIBinaryPath, "md", "apply",
-			}, " ")
-			cmd := exec.Command("bash", "-c", args)
-			cmd.Dir = dirName
-			session, err := Start(
-				cmd,
-				NewPrefixedWriter(testutil.DebugOutPrefix, GinkgoWriter),
-				NewPrefixedWriter(testutil.DebugErrPrefix, GinkgoWriter),
-			)
-			Expect(err).To(BeNil())
-			Eventually(session, 60*40).Should(Exit(0))
-
-			session = testutil.Hasura(testutil.CmdOpts{
-				Args:             []string{"metadata", "export", "--output", "yaml"},
-				WorkingDirectory: dirName,
-			})
-			Eventually(session, 60*40).Should(Exit(0))
-			out := session.Wait().Out.Contents()
-			Expect(isYAML(out)).To(BeTrue())
-			Eventually(string(out)).Should(ContainSubstring("max_connections: 51"))
 		})
 	})
 })