+++ CHANGELOG.md
@@ -69,6 +69,7 @@ $ hasura metadata export -o json
 - cli: add warning for incompatible pro plugin version
 - cli: add new sub command `delete` to `hasura migrate`
 - cli: fix bug in migrate squash due to empty down file (#3897)
+- cli: fix bug with metadata apply on some CI environments (#6987)
 
 ## v2.0.0-alpha.10